# Conflicts:
#	src/main/java/com/njmsita/exam/manager/controller/QuestionBankManagerController.java
#	src/main/java/com/njmsita/exam/manager/service/ebi/TopicEbi.java

+++ src/main/java/com/njmsita/exam/manager/dao/dao/QuestionDao.java
@@ -23,5 +23,5 @@ public interface QuestionDao extends BaseDao<QuestionVo>
      * @param id    题型id
      * @return
      */
-    public List<QuestionVo> getByQuestionType(Byte id);
+    public List<QuestionVo> getByQuestionType(Integer id);
 }
