+++ src/main/java/com/njmsita/exam/manager/model/QuestionTypeVo.java
@@ -16,7 +16,7 @@ public class QuestionTypeVo
 {
 
     @NotNull(message = "{id.notempty}",groups = {EditGroup.class})
-    private byte id;
+    private Integer id;
     @NotEmpty(message = "{name.notempty}",groups = {AddGroup.class, EditGroup.class})
     private String name;
     @NotNull(message = "{score.notempty}",groups = {AddGroup.class, EditGroup.class})
@@ -24,12 +24,12 @@ public class QuestionTypeVo
 
     @Id
     @Column(name = "id")
-    public byte getId()
+    public Integer getId()
     {
         return id;
     }
 
-    public void setId(byte id)
+    public void setId(Integer id)
     {
         this.id = id;
     }
